SAGE-1622: pipeline to build the RPI SD flash image
Launch a docker container that downloads the base RPI image,
updates the eeprom tool to include the desired firmware version, and
applies the Waggle customizations. The produced compressed (xz) SD
card image file contains automatic firmware update files in the
/boot partition that are automatically applied on first boot by
the RPI bootloader to configure the RPI for PxE (i.e. net) boot mode.
The /root partition contains the base RPI image that is used as
a backup OS in the event that PxE booting fails.

Boot configuration summary
 - netboot is tried first
 - after 5x failures the SD card (backup) OS is booted